migration_steps/acquire_target_ids/requirements.txt
sqlalchemy
psycopg2
pandas
xlrd
python-dotenv
